Move protocol logic out of StateMachine
